# fix: omit data-race panic on `GC()`